requirements.txt
alabaster==0.7.10
astroid==1.6.1
attrs==17.4.0
autopep8==1.3.4
Babel==2.5.3
certifi==2018.1.18
chardet==3.0.4
coverage==4.5.1
docutils==0.14
flake8==3.5.0
hypothesis==3.44.26
idna==2.6
imagesize==0.7.1
isort==4.3.4
Jinja2==2.10
lazy-object-proxy==1.3.1
MarkupSafe==1.0
mccabe==0.6.1
mypy==0.560
numpy==1.14.0
packaging==16.8
Pillow==5.0.0
pluggy==0.6.0
psutil==5.4.3
py==1.5.2
py-cpuinfo==3.3.0
pycodestyle==2.3.1
pyflakes==1.6.0
Pygments==2.2.0
pylint==1.8.2
pyparsing==2.2.0
pytest==3.4.0
pytest-benchmark==3.1.1
pytest-cov==2.5.1
pytz==2018.3
requests==2.18.4
six==1.11.0
snowballstemmer==1.2.1
Sphinx==1.7.0
sphinxcontrib-websupport==1.0.1
typed-ast==1.1.0
urllib3==1.22
wrapt==1.10.11
